requirements.txt
Flask==2.3.2
Flask-WTF==1.2.1
WTForms==3.0.1
Werkzeug==3.0.3
SQLAlchemy==2.0.22
Flask-SQLAlchemy==3.0.4
pandas==2.2.2
matplotlib==3.8.0
python-dateutil==2.8.2
pytest==7.4.0
